requirements.txt
streamlit==0.88.0
pandas==1.1.3
numpy==1.19.2
pillow==8.0.1
scikit-learn==0.23.2
pickleshare==0.7.5
keras==2.6.0
tensorflow==2.6.0
streamlit-player==0.1.5
